[cute] move tcgen05 support envelope to a dependency-free module + trim it
Two low review findings on the envelope refactor.

Import fragility: enforce_dot_requirements (helion/language) imported the
envelope from cute_mma.py, crossing into the heavy CuTe codegen module and
closing a latent cycle (cute_mma -> aux_tensor -> matmul_ops; it only worked
because the import was local). Move Tcgen05MatmulEnvelope +
tcgen05_unsupported_reason to a new dependency-free
helion/_compiler/cute/tcgen05_support.py, imported at module level by both
cute_mma and matmul_ops -- no cycle, no cross into codegen.

"Single source of truth" was slightly overstated: the envelope carried two
fields (uses_tma_ab_pipeline, k_tile_count) that no rule used and that the
autotune caller could only default, so a future rule using them could
silently diverge from codegen. Drop them (YAGNI) and remove all field
defaults so every field is required, used by a rule, and populated with
accurate values by BOTH callers -- codegen from the resolved config, autotune
from the candidate edge-2-CTA config (cluster_n=1/persistent/all-partial are
that candidate's real values, not placeholders, since the edge-search gate
requires m+n edges and a k tail). Add a field only with the rule that needs
it, updating both call sites.

Behavior unchanged (adversarial sweep still 0 silent-wrong). Full cute suite
174 passed; dot_requirements 51 passed.
